+++ client/pages/[username]/[slug]/index.tsx
@@ -127,7 +127,7 @@ const Preview: NextPage<Props> = ({ username, slug, resume: initialData }) => {
       </div>
 
       <p className={styles.footer}>
-        Made with <Link href="/">Reactive Resume</Link>
+        Made with <Link href="/">TivlotCV</Link>
       </p>
     </div>
   );