+++ client/pages/r/[shortId].tsx
@@ -100,7 +100,7 @@ const Preview: NextPage<Props> = ({ shortId }) => {
       </div>
 
       <p className={styles.footer}>
-        Made with <Link href="/">Reactive Resume</Link>
+        Made with <Link href="/">TivlotCV</Link>
       </p>
     </div>
   );